requirements.txt
pyYAML
slacker
